requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-Migrate
python-dotenv
boto3
cryptography
psycopg2-binary
PyMySQL
gunicorn
celery==5.3.6  # Add this
redis==5.0.1   # Add this
celery[redis]==5.3.6
flower